+++ index.d.ts
@@ -67,10 +67,6 @@ declare type TokenBalance = {
     reserved: TReservedBalance;
     frozen: TFreeBalance;
 };
-declare type Reward = {
-    notYetClaimed: BN;
-    toBeClaimed: BN;
-};
 
 interface MangataEventData {
     lookupName: string;
@@ -105,7 +101,7 @@ declare type XcmTxOptions = Partial<Omit<TxOptions, "statusCallback" | "extrinsi
  */
 declare class Mangata {
     private api;
-    private uri;
+    private urls;
     private static instanceMap;
     /**
      * The Mangata's constructor is private to prevent direct
@@ -120,7 +116,7 @@ declare class Mangata {
     /**
      * The static method that controls the access to the Mangata instance.
      */
-    static getInstance(uri: string): Mangata;
+    static getInstance(urls: string[]): Mangata;
     /**
      * Api instance of the connected node
      */
@@ -128,7 +124,7 @@ declare class Mangata {
     /**
      * Uri of the connected node
      */
-    getUri(): string;
+    getUrls(): string[];
     /**
      * Wait for the new block
      */
@@ -447,4 +443,4 @@ declare const signTx: (api: ApiPromise, tx: SubmittableExtrinsic<"promise">, acc
     extrinsicStatus: (events: MangataGenericEvent[]) => void;
 }> | undefined) => Promise<MangataGenericEvent[]>;
 
-export { AssetInfo, BIG_BILLION, BIG_HUNDRED, BIG_HUNDRED_BILLIONS, BIG_HUNDRED_MILLIONS, BIG_HUNDRED_THOUSAND, BIG_MILLION, BIG_ONE, BIG_TEN, BIG_TEN_BILLIONS, BIG_TEN_MILLIONS, BIG_TEN_THOUSAND, BIG_THOUSAND, BIG_TRILLION, BIG_ZERO, BN_BILLION, BN_DIV_NUMERATOR_MULTIPLIER, BN_DIV_NUMERATOR_MULTIPLIER_DECIMALS, BN_HUNDRED, BN_HUNDRED_BILLIONS, BN_HUNDRED_MILLIONS, BN_HUNDRED_THOUSAND, BN_MILLION, BN_ONE, BN_TEN, BN_TEN_BILLIONS, BN_TEN_MILLIONS, BN_TEN_THOUSAND, BN_THOUSAND, BN_TRILLION, BN_ZERO, Mangata, MangataEventData, MangataGenericEvent, MangataHelpers, Reward, TBalances, TFreeBalance, TFrozenBalance, TMainTokens, TPool, TPoolWithRatio, TPoolWithShare, TReservedBalance, TToken, TTokenAddress, TTokenId, TTokenInfo, TTokenMainInfo, TTokenName, TTokenSymbol, TTokens, TokenBalance, TxOptions, XcmTxOptions, fromBN, signTx, toBN, toFixed };
+export { AssetInfo, BIG_BILLION, BIG_HUNDRED, BIG_HUNDRED_BILLIONS, BIG_HUNDRED_MILLIONS, BIG_HUNDRED_THOUSAND, BIG_MILLION, BIG_ONE, BIG_TEN, BIG_TEN_BILLIONS, BIG_TEN_MILLIONS, BIG_TEN_THOUSAND, BIG_THOUSAND, BIG_TRILLION, BIG_ZERO, BN_BILLION, BN_DIV_NUMERATOR_MULTIPLIER, BN_DIV_NUMERATOR_MULTIPLIER_DECIMALS, BN_HUNDRED, BN_HUNDRED_BILLIONS, BN_HUNDRED_MILLIONS, BN_HUNDRED_THOUSAND, BN_MILLION, BN_ONE, BN_TEN, BN_TEN_BILLIONS, BN_TEN_MILLIONS, BN_TEN_THOUSAND, BN_THOUSAND, BN_TRILLION, BN_ZERO, Mangata, MangataEventData, MangataGenericEvent, MangataHelpers, TBalances, TFreeBalance, TFrozenBalance, TMainTokens, TPool, TPoolWithRatio, TPoolWithShare, TReservedBalance, TToken, TTokenAddress, TTokenId, TTokenInfo, TTokenMainInfo, TTokenName, TTokenSymbol, TTokens, TokenBalance, TxOptions, XcmTxOptions, fromBN, signTx, toBN, toFixed };