+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.1](https://github.com/mangata-finance/mangata-SDK/compare/v1.0.0...v1.0.1) (2022-06-14)
+
 ## [1.0.0](https://github.com/mangata-finance/mangata-SDK/compare/v0.5.5...v1.0.0) (2022-06-14)
 
 - official release of Mangta SDK